pyproject.toml
[project]
name = "jurnalyuga"
version = "0.1.0"
description = ""
authors = [
    {name = "NikitaErmakov",email = "ermafin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "uvicorn[standart] (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.1.2,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.3"
aiosqlite = "^0.21.0"
trio = "^0.29.0"

